eoCounter? eoEasyEA -- made it copyable again eoEvalFunc -- added specialized eoEvalFuncCounter eoEvolutionStrategy -- nothing much eoGenContinue -- nothing eoPop -- fixed nth_element_fitness eoBitOp -- fixed error in xover eoFileMonitor -- now appends always eoParam -- worked around memory leak in MSC's strstream eoParser -- changed -pconfig_file to @config_file eoParser -- added messages instead of exception when required param is missing eoStat -- added eoDistanceStat t-eoFunctor -- don't know

+++ eo/src/eoGenContinue.h
@@ -36,13 +36,13 @@ class eoGenContinue: public eoContinue<EOT>
 public:
 
 	/// Ctor for setting a
-	eoGenContinue( unsigned _totalGens)
+	eoGenContinue( unsigned long _totalGens)
 	  : repTotalGenerations( _totalGens ), 
       thisGenerationPlaceHolder(0),
 	  thisGeneration(thisGenerationPlaceHolder){};
 	
     /// Ctor for enabling the save/load the no. of generations counted
-	eoGenContinue( unsigned _totalGens, unsigned& _currentGen)
+	eoGenContinue( unsigned long _totalGens, unsigned long& _currentGen)
 	  : repTotalGenerations( _totalGens ), 
       thisGenerationPlaceHolder(0),
 	  thisGeneration(_currentGen){};
@@ -57,21 +57,21 @@ public:
 
 	/** Sets the number of generations to reach 
 	    and sets the current generation to 0 (the begin)*/
-	virtual void totalGenerations( unsigned _tg ) { 
+	virtual void totalGenerations( unsigned long _tg ) { 
 	  repTotalGenerations = _tg; 
 	  thisGeneration = 0;
 	};
 
 	/** Returns the number of generations to reach*/
-	virtual unsigned totalGenerations( ) 
+	virtual unsigned long totalGenerations( ) 
     {  
 	  return repTotalGenerations; 
 	};
     
 private:
-	unsigned repTotalGenerations;
-    unsigned thisGenerationPlaceHolder;
-    unsigned& thisGeneration;
+	unsigned long repTotalGenerations;
+    unsigned long thisGenerationPlaceHolder;
+    unsigned long& thisGeneration;
 };
 
 #endif
